+++ apps/frontend/src/app/3_organisms/ConversionsHistoryFrame/components/BobConversionHistoryFrame/BobConversionsHistoryFrame.tsx
@@ -16,6 +16,8 @@ import {
   NotificationType,
 } from '@sovryn/ui';
 
+import { BOB_CHAIN_ID } from '../../../../../config/chains';
+
 import {
   DEFAULT_HISTORY_FRAME_PAGE_SIZE,
   EXPORT_RECORD_LIMIT,
@@ -33,9 +35,11 @@ import {
   useGetSwapHistoryLazyQuery,
 } from '../../../../../utils/graphql/bob/generated';
 import { dateFormat } from '../../../../../utils/helpers';
-import { fromWei } from '../../../../../utils/math';
 import { BaseConversionsHistoryFrame } from '../BaseConversionsHistoryFrame/BaseConversionsHistoryFrame';
-import { COLUMNS_CONFIG } from './BobConversionsHistoryFrame.constants';
+import {
+  COLUMNS_CONFIG,
+  getConversionAmount,
+} from './BobConversionsHistoryFrame.constants';
 import { generateRowTitle } from './BobConversionsHistoryFrame.utils';
 import { useGetBobConversionsHistory } from './hooks/useGetBobConversionsHistory';
 
@@ -101,18 +105,14 @@ export const BobConversionsHistoryFrame: React.FC<PropsWithChildren> = ({
         tx.qty === '0'
           ? t(translations.conversionsHistory.swapMultihop)
           : t(translations.conversionsHistory.swap),
-      sent: tx.inBaseQty
-        ? Math.abs(Number(fromWei(tx.baseFlow)))
-        : Math.abs(Number(fromWei(tx.quoteFlow))),
-      sentToken: tx.inBaseQty
-        ? getTokenDisplayNameByAddress(tx.pool.base, getCurrentChain())
-        : getTokenDisplayNameByAddress(tx.pool.quote, getCurrentChain()),
+      sent: getConversionAmount(tx as Swap, true).amount.toString(),
+      setToken: tx.inBaseQty
+        ? getTokenDisplayNameByAddress(tx.pool.base, BOB_CHAIN_ID)
+        : getTokenDisplayNameByAddress(tx.pool.quote, BOB_CHAIN_ID),
       receivedToken: tx.inBaseQty
-        ? getTokenDisplayNameByAddress(tx.pool.quote, getCurrentChain())
-        : getTokenDisplayNameByAddress(tx.pool.base, getCurrentChain()),
-      received: tx.inBaseQty
-        ? Math.abs(Number(fromWei(tx.quoteFlow)))
-        : Math.abs(Number(fromWei(tx.baseFlow))),
+        ? getTokenDisplayNameByAddress(tx.pool.quote, BOB_CHAIN_ID)
+        : getTokenDisplayNameByAddress(tx.pool.base, BOB_CHAIN_ID),
+      received: getConversionAmount(tx as Swap, false).amount.toString(),
       TXID: tx.transactionHash,
     }));
   }, [account, addNotification, getConversions]);