+++ apps/frontend/src/app/5_pages/ClaimPowaPage/ClaimPowaPage.constants.tsx
@@ -1,0 +1,3 @@
+export const TELEGRAM_POWA_URL = 'https://t.me/SovrynBitcoin';
+export const ABOUT_POWA_URL = 'https://wiki.sovryn.com/en/campaign/powa';
+export const SUBMIT_WALLET_URL = 'https://gleam.io/EK36C/influence-powa';