+++ apps/frontend/src/constants/infrastructure/rsk.ts
@@ -2,14 +2,8 @@ import { Environments } from '../../types/global';
 
 export const RSK = {
   rpc: {
-    [Environments.Mainnet]: [
-      'https://rsk-live.sovryn.app/rpc',
-      'https://public-node.rsk.co',
-    ],
-    [Environments.Testnet]: [
-      'https://testnet.sovryn.app/rpc',
-      'https://public-node.testnet.rsk.co',
-    ],
+    [Environments.Mainnet]: ['https://rsk-live.sovryn.app/rpc'],
+    [Environments.Testnet]: ['https://testnet.sovryn.app/rpc'],
   },
   publicRpc: {
     [Environments.Mainnet]: 'https://mainnet.sovryn.app/rpc',