+++ packages/sdk/src/_tests/swaps/smart-router.test.ts
@@ -91,7 +91,7 @@ describe('SmartRouter', () => {
     });
 
     it('get cheapest swap route for DLLR -> BTC', async () => {
-      const amount = ethers.utils.parseEther('20');
+      const amount = ethers.utils.parseEther('0.01');
       await expect(
         router.getBestQuote(chainId, dllr, btc, amount),
       ).resolves.toMatchObject({
@@ -101,7 +101,7 @@ describe('SmartRouter', () => {
     });
 
     it('get cheapest swap route for BTC -> DLLR', async () => {
-      const amount = ethers.utils.parseEther('20');
+      const amount = ethers.utils.parseEther('0.01');
       await expect(
         router.getBestQuote(chainId, dllr, btc, amount),
       ).resolves.toMatchObject({
@@ -117,12 +117,12 @@ describe('SmartRouter', () => {
     });
 
     it('returns all available entries', async () => {
-      await expect(router.getEntries(chainId)).resolves.toHaveLength(14);
+      await expect(router.getEntries(chainId)).resolves.toHaveLength(15);
     });
 
     it('returns all available destinations for entry token', async () => {
       await expect(router.getDestination(chainId, sov)).resolves.toHaveLength(
-        12,
+        13,
       );
     });