+++ packages/sdk/src/swaps/smart-router/routes/amm-swap-route.ts
@@ -110,6 +110,7 @@ export const ammSwapRoute: SwapRouteFunction = (
           'XUSD',
           'MYNT',
           'BPRO',
+          'POWA',
         ];
 
         const contracts = (